Various changes to runkit for PHP5 -> PHP7 migration; disable import, constants, properties
The following php runkit features are not yet supported
(The code still needs to be modified to work in PHP7)
-	`runkit_sandbox_*`, using `./configure --disable-runkit-sandbox --disable-runkit-super
-	runkit superglobal
-	`runkit_import_*, with the new (undefined) macro
	PHP_RUNKIT_MANIPULATION_IMPORT
	(uncomment #define PHP_RUNKIT_MANIPULATION_IMPORT to try
	compiling it)
-	`runkit_class_**` and `runkit_property_**`,
	(uncomment #define PHP_RUNKIT_MANIPULATION_PROPERTIES to try
	compiling it)

The znode_op structure changed.
TODO: Change runkit to account for this.

Make feature constant persistent (is this right?)

Start replacing zend_hash_apply with macros and functions.
-	If a raw pointer is added to the hash, then the
	zend_apply_func_t no longer gets that pointer.
	Instead, it gets a zval with `Z_TYPE_P(zv) == IS_PTR`
	that can be fetched with `Z_PTR_T(zv)`

Remove more `#if` conditional code for PHP versions below 5.6